add overclock + disable drag / deletion when input is targeted

+++ src/app/components/item/item.component.ts
@@ -1,6 +1,7 @@
-import { Component, HostListener, inject, OnInit } from "@angular/core";
+import { Component, computed, HostListener, inject, OnInit } from "@angular/core";
 import { MatCardModule } from "@angular/material/card";
 import { MatCheckboxChange, MatCheckboxModule } from "@angular/material/checkbox";
+import { MatChipsModule } from "@angular/material/chips";
 import { MatDividerModule } from "@angular/material/divider";
 import { MatFormFieldModule } from "@angular/material/form-field";
 import { MatInputModule } from "@angular/material/input";
@@ -14,16 +15,46 @@ import { EntitiesService } from "../../services/entities.service";
   templateUrl: "./item.component.html",
   styleUrl: "./item.component.scss",
   standalone: true,
-  imports: [EditorModule, MatCardModule, MatDividerModule, MatInputModule, MatFormFieldModule, MatCheckboxModule],
+  imports: [
+    EditorModule,
+    MatCardModule,
+    MatDividerModule,
+    MatInputModule,
+    MatFormFieldModule,
+    MatCheckboxModule,
+    MatChipsModule,
+  ],
 })
 export class ItemComponent extends CustomDynamicNodeComponent<ItemData> implements OnInit {
   private dataService = inject(DataService);
   private entitiesService = inject(EntitiesService);
 
+  public overclock = computed(() => {
+    return this.data() !== undefined ? this.data()!.overclock : 100;
+  });
+
   public constructedCheckChange(event: MatCheckboxChange) {
     this.data.set({
       ...this.data(),
       constructed: event.checked,
+    } as ItemData);
+  }
+
+  public overclockChange(event: Event) {
+    const target = event.target as HTMLInputElement;
+    let value = target.valueAsNumber;
+
+    if (value < 1) {
+      value = 1;
+      target.valueAsNumber = 1;
+    } else if (value > 250) {
+      value = 250;
+      target.valueAsNumber = 250;
+    }
+
+    this.data.set({
+      ...this.data(),
+      overclock: target.valueAsNumber,
     } as ItemData);
   }
 
@@ -34,9 +65,12 @@ export class ItemComponent extends CustomDynamicNodeComponent<ItemData> implemen
     this.recipe = this.dataService.recipes.find((el) => el.name === this.data()?.name)!;
   }
 
-  @HostListener("window:keydown.delete")
-  @HostListener("window:keydown.backspace")
-  public deleteNode() {
+  @HostListener("window:keydown.delete", ["$event"])
+  @HostListener("window:keydown.backspace", ["$event"])
+  public deleteNode(event: KeyboardEvent) {
+    if (event.target instanceof HTMLInputElement) {
+      return;
+    }
     if (this.selected()) {
       this.entitiesService.deleteNode(this.node.id);
     }
@@ -46,7 +80,5 @@ export class ItemComponent extends CustomDynamicNodeComponent<ItemData> implemen
     return key ? this.dataService.getBgForItem(key) : "";
   }
 
-  public getIcon(key: string) {
-    
-  }
+  public getIcon(key: string) {}
 }
